Update contains operators to handle strings properly (#56)
